oms_hibernate/oms/src/main/java/com/sapient/oms/entity/InventoryKey.java
package com.sapient.oms.entity;

import java.io.Serializable;

import javax.persistence.Column;
import javax.persistence.Embeddable;

import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@Embeddable
@NoArgsConstructor
@AllArgsConstructor
public class InventoryKey implements Serializable {

    @Column(name = "store_id")
    int storeid;

    @Column(name = "product_id")
    int productid;

    public int getStoreid() {
        return storeid;
    }

    public void setStoreid(int storeid) {
        this.storeid = storeid;
    }

    public int getProductid() {
        return productid;
    }

    public void setProductid(int productid) {
        this.productid = productid;
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + productid;
        result = prime * result + storeid;
        return result;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj)
            return true;
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        InventoryKey other = (InventoryKey) obj;
        if (productid != other.productid)
            return false;
        if (storeid != other.storeid)
            return false;
        return true;
    }

}
